implemented the business(service) layer

+++ src/main/java/ke/co/turbosoft/tt/repo/CompanyRepo.java
@@ -4,6 +4,6 @@ import ke.co.turbosoft.tt.entity.Company;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepo extends JpaRepository<Company, Long> {
+public interface CompanyRepo extends JpaRepository<Company, Integer> {
 
 }
