+++ src/main/java/ke/co/turbosoft/tt/repo/ProjectRepo.java
@@ -4,6 +4,6 @@ import ke.co.turbosoft.tt.entity.Project;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepo extends JpaRepository<Project, Long> {
+public interface ProjectRepo extends JpaRepository<Project, Integer> {
 
 }
