+++ src/main/java/ke/co/turbosoft/tt/repo/TaskRepo.java
@@ -4,6 +4,6 @@ import ke.co.turbosoft.tt.entity.Task;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRepo extends JpaRepository<Task, Long> {
+public interface TaskRepo extends JpaRepository<Task, Integer> {
 
 }
